requirement.txt
django
qrcode
pillow
myqr
uuid
imageio